+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './index/index.js'
-export * from './messages/MessagesBuilder.js'
+export * from './messages/index.js'
+export * from './step-documents/index.js'
 export * from './tree-sitter/index.js'
