+++ src/index/jsSearchIndex.ts
@@ -1,6 +1,6 @@
 import { Search } from 'js-search'
 
-import { StepDocument } from './step-documents/types.js'
+import { StepDocument } from '../step-documents/types.js'
 import { Index } from './types.js'
 
 type Doc = {
