+++ src/index/types.ts
@@ -4,6 +4,6 @@
  * @param text a text to search for
  * @return results in the form of step documents
  */
-import { StepDocument } from './step-documents/types.js'
+import { StepDocument } from '../step-documents/types.js'
 
 export type Index = (text: string) => readonly StepDocument[]
