+++ src/service/snippet/lspCompletionSnippet.ts
@@ -3,7 +3,7 @@
  *
  * @param expression the expression to generate the snippet from
  */
-import { StepSegments } from '../../index/step-documents/types.js'
+import { StepSegments } from '../../step-documents/types.js'
 
 export function lspCompletionSnippet(stepSegments: StepSegments): string {
   let n = 1
